requirements.txt
kafka-python
slackweb
Fabric3